Address final RMS normalize review comments
Agent-Logs-Url: https://github.com/jianan-gu/sglang/sessions/9ac8b9c0-a528-4f01-9bbd-067cd1cc0005

+++ python/sglang/srt/layers/layernorm.py
@@ -44,7 +44,7 @@ def rms_normalize_native(
     normalized = x_f32 * torch.rsqrt(variance + eps)
 
     if weight is not None:
-        normalized = normalized * weight
+        normalized = normalized * weight.to(dtype=normalized.dtype)
 
     x.copy_(normalized.to(dtype=x.dtype))
     return x
